+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.O365OrgSettings.Tests.ps1
@@ -39,10 +39,38 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             Mock -CommandName Get-MgServicePrincipal -MockWith {
             }
 
+            Mock -CommandName Set-DefaultTenantBriefingConfig -MockWith {
+            }
+
+            Mock -CommandName Set-DefaultTenantMyAnalyticsFeatureConfig -MockWith {
+            }
             Mock -CommandName Invoke-MgGraphRequest -MockWith {
                 return @{
                     "@odata.type"         = "#microsoft.graph.adminReportSettings"
                     displayConcealedNames = $true
+                }
+            }
+
+            Mock -CommandName Get-M365DSCOrgSettingsInstallationOptions -MockWith {
+                return @{
+                    '@odata.context' = 'https://graph.microsoft.com/beta/$metadata#admin/microsoft365Apps/installationOptions/$entity'
+                    updateChannel = 'current'
+                    appsForMac = @{
+                        isSkypeForBusinessEnabled = $True
+                        isMicrosoft365AppsEnabled = $true
+                    }
+                    appsForWindows = @{
+                        isVisioEnabled            = $True
+                        isSkypeForBusinessEnabled = $False
+                        isMicrosoft365AppsEnabled = $true
+                        isProjectEnabled          = $true
+                    }
+                }
+            }
+
+            Mock -CommandName Get-M365DSCO365OrgSettingsPlannerConfig -MockWith {
+                return @{
+                    allowCalendarSharing = $false
                 }
             }
         }
@@ -54,7 +82,15 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     AdminCenterReportDisplayConcealedNames     = $True;
                     IsSingleInstance                           = 'Yes'
                     M365WebEnableUsersToOpenFilesFrom3PStorage = $False;
-                    Ensure                                     = 'Present'
+                    InstallationOptionsAppsForMac              = @('isSkypeForBusinessEnabled', 'isMicrosoft365AppsEnabled')
+                    InstallationOptionsAppsForWindows          = @('isVisioEnabled', 'isMicrosoft365AppsEnabled', 'isProjectEnabled')
+                    InstallationOptionsUpdateChannel           = 'current'
+                    MicrosoftVivaBriefingEmail                   = $True
+                    VivaInsightsWebExperience                    = $true
+                    VivaInsightsDigestEmail                      = $true
+                    VivaInsightsOutlookAddInAndInlineSuggestions = $true
+                    VivaInsightsScheduleSendSuggestions          = $true
+                    PlannerAllowCalendarSharing                = $False
                     Credential                                 = $Credential
                 }
 
@@ -63,10 +99,24 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         AccountEnabled = $False
                     }
                 }
+
+                Mock -CommandName Get-DefaultTenantBriefingConfig -MockWith {
+                    return @{
+                        IsEnabledByDefault = 'opt-in'
+                    }
+                }
+
+                Mock -CommandName Get-DefaultTenantMyAnalyticsFeatureConfig -MockWith {
+                    return @{
+                        IsDashboardEnabled    = $true
+                        IsDigestEmailEnabled  = $true
+                        IsAddInEnabled        = $true
+                        IsScheduleSendEnabled = $true
+                    }
+                }
             }
 
             It 'Should return Present from the Get method' {
-                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
                 (Get-TargetResource @testParams).M365WebEnableUsersToOpenFilesFrom3PStorage | Should -Be $False
             }
 
@@ -79,11 +129,16 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name 'When Org Settings NOT in the Desired State' -Fixture {
             BeforeAll {
                 $testParams = @{
-                    AdminCenterReportDisplayConcealedNames     = $True;
-                    IsSingleInstance                           = 'Yes'
-                    M365WebEnableUsersToOpenFilesFrom3PStorage = $True;
-                    Ensure                                     = 'Present'
-                    Credential                                 = $Credential
+                    IsSingleInstance                             = 'Yes'
+                    AdminCenterReportDisplayConcealedNames       = $True;
+                    M365WebEnableUsersToOpenFilesFrom3PStorage   = $True
+                    MicrosoftVivaBriefingEmail                   = $True
+                    VivaInsightsWebExperience                    = $true
+                    VivaInsightsDigestEmail                      = $true
+                    VivaInsightsOutlookAddInAndInlineSuggestions = $true
+                    VivaInsightsScheduleSendSuggestions          = $true
+                    Ensure                                       = 'Present'
+                    Credential                                   = $Credential
                 }
 
                 Mock -CommandName Get-MgServicePrincipal -MockWith {
@@ -91,10 +146,24 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         AccountEnabled = $False
                     }
                 }
+
+                Mock -CommandName Get-DefaultTenantBriefingConfig -MockWith {
+                    return @{
+                        IsEnabledByDefault = 'opt-in'
+                    }
+                }
+
+                Mock -CommandName Get-DefaultTenantMyAnalyticsFeatureConfig -MockWith {
+                    return @{
+                        IsDashboardEnabled    = $true
+                        IsDigestEmailEnabled  = $true
+                        IsAddInEnabled        = $true
+                        IsScheduleSendEnabled = $true
+                    }
+                }
             }
 
             It 'Should return Present from the Get method' {
-                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
                 (Get-TargetResource @testParams).M365WebEnableUsersToOpenFilesFrom3PStorage | Should -Be $False
             }
 
@@ -122,6 +191,22 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         AccountEnabled = $False
                     }
                 }
+
+                Mock -CommandName Get-DefaultTenantBriefingConfig -MockWith {
+                    return @{
+                        IsEnabledByDefault = 'opt-in'
+                    }
+                }
+
+                Mock -CommandName Get-DefaultTenantMyAnalyticsFeatureConfig -MockWith {
+                    return @{
+                        IsDashboardEnabled    = $true
+                        IsDigestEmailEnabled  = $true
+                        IsAddInEnabled        = $true
+                        IsScheduleSendEnabled = $true
+                    }
+                }
+
                 $result = Export-TargetResource @testParams
                 $result | Should -Not -BeNullOrEmpty
             }